+++ protocol_wire/src/SellerTerms.cpp
@@ -19,7 +19,7 @@ namespace protocol_wire {
         : SellerTerms(0,0,0,0,0) {
     }
 
-    SellerTerms::SellerTerms(quint64 price, quint32 lock, quint32 maxSellers, quint64 minContractFeePerKb, quint64 settlementFee)
+    SellerTerms::SellerTerms(quint64 price, quint16 lock, quint32 maxSellers, quint64 minContractFeePerKb, quint64 settlementFee)
         : _minPrice(price)
         , _minLock(lock)
         , _maxSellers(maxSellers)
@@ -84,11 +84,11 @@ namespace protocol_wire {
         _minPrice = price;
     }
 
-    quint32 SellerTerms::minLock() const {
+    quint16 SellerTerms::minLock() const {
         return _minLock;
     }
 
-    void SellerTerms::setMinLock(quint32 minLock) {
+    void SellerTerms::setMinLock(quint16 minLock) {
         _minLock = minLock;
     }
